refactoring post dashes and previews
