+++ client/src/App.js
@@ -4,7 +4,7 @@ import Landing from './components/Landing';
 import {BlogDash, ForumDash} from './components/PostComponents/PostDashes';
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import UserContextProvider from "./contexts/UserContextProvider";
-import Navbar from './components/Navbar';
+import NavBar from './components/NavBar';
 import Signup from './components/Signup';
 import Footer from './components/Footer';
 
@@ -13,7 +13,7 @@ function App() {
       <Router>
         <UserContextProvider>
           <div id="bodyColor"></div>
-          <Navbar />
+          <NavBar />
           <Switch>
             <Route path="/" exact component={Landing}></Route>
             <Route path="/login" component={Login}></Route>
